chore: move format_panic_data to starknet-api
