+++ crates/blockifier/src/execution/syscalls/syscall_tests/call_contract.rs
@@ -4,6 +4,7 @@ use std::sync::Arc;
 use itertools::Itertools;
 use pretty_assertions::assert_eq;
 use rstest::rstest;
+use starknet_api::execution_utils::format_panic_data;
 use starknet_api::felt;
 use starknet_api::transaction::Calldata;
 use test_case::test_case;
@@ -14,7 +15,6 @@ use crate::context::ChainInfo;
 use crate::execution::call_info::{CallExecution, Retdata};
 use crate::execution::contract_class::TrackedResource;
 use crate::execution::entry_point::CallEntryPoint;
-use crate::execution::execution_utils::format_panic_data;
 use crate::retdata;
 use crate::state::state_api::StateReader;
 use crate::test_utils::contracts::FeatureContract;
